[#12] Fix SpringLiquibase and Liquibase CLI interop
By default, SpringLiquibase requires a 'classpath:' prefix for the change log filename while the Liquibase CLI does not. This leads to a filename mismatch in the databasechangelog table of the database. While Spring strips the prefix when checking whether or not a change set has been applied already, Liquibase does not. This leads to interoperability issues when using both SpringLiquibase and the Liquibase CLI.

See https://liquibase.jira.com/browse/CORE-2766 for more information.