[#1] Set PostgreSQL password from random to fixed
This fixes an issue when the user applied changes: The habit and track chart was always redeployed because the Helm diff plugin recognized that the password changed from last deployment.

Another issue was caused by helmfile not deleting persistent volumes and persistent volume claims on destroy. Because of this, the actual password and the password stored in the secret were suddenly out of sync.