Enable OIDC for default deployment
The cluster local domain `http://keycloak-http.hc-keycloak.svc.cluster.local`
is used as host for authorization and token endpoint. Subsequently the
Keycloak integration requires that the cluster local domain is reachable from the
user agent (e.g. by running `minikube tunnel`).

To fix this we must separate the configuration of the authorization and token
endpoint.