[#19] Fix locust SSL connection error when generating load
Locust is deployed inside of the k8s cluster where the domain
habitcentric.demo can not be resolved.

Host aliases can not be used because the resolved IP must match the IP
of the Istio ingress gateway.

To reach the Istio ingress gateway anyways, Locust uses the FQDN
istio-ingressgateway.istio-system.svc.cluster.local and sets the Host
header to habitcentric.demo. This does not work when SSL is enabled for
the gateway because the SNI check during the TLS handshake fails. To fix
this, we add the FQDN of the ingress gateway to the TLS certificate of
the gateway with this change.