Remove fault injection
The fault injection via traffic split cannot be used with
resilience settings, since the faulty service has a
separate Kubernetes service (and thus needs a
separate ServiceProfile). We can test retries to the
report service this way, but we decided that testing
retries on the report service with a faulty track service
is not straight forward enough.