Fix Kiali not detecting namespaces after Kiali CR deployment
By default, the Kiali operator only detects namespaces when running a
reconciliation. If namespaces are created after the reconciliation,
the Kiali operator will never have created the necessary roles for Kiali
to see them.

A workaround is setting the accessible namespaces to '**'. In this case,
the operator creates a cluster role instead of separate roles for each
accessible namespace. This is why `clusterRoleCreater` must be set to
`true`.