Deploy Kiali using Helm chart instead of operator
Helm chart makes more sense for our demo setup because uninstalling
deployments using the operator is always a hassle. When the operator is
used, you have to uninstall the CR first, wait until the operator
removed the deployment, and then remove the operator. With Helm you can
just uninstall the releases and you are done.

Installing Kiali using Helm is also more in line with our Istio and
habitcentric deployment.