Revert "ci: fix skipped checks due to path filter (#21)"
This reverts commit a0f8b09d8a965e5ff113582cbf32636c75adf30f.

The idea of the the dummy checks (recommended by github) in combination
with the monorepo is too complicated.
The required logic to make the checks work properly with a monorepo is
not available with native github actions and would require us to add a
lot of complicated logic around the pipeline...

So we are just dumping the required checks for PRs.