fix: assign user role to users logged in via oauth
Spring Boot 3 changed the default authority of users logged in with
OAuth from `ROLE_USER` to `OAUTH2_USER`. This broke our
authorization, because it was based on the USER role. This commit
reintroduces the USER role for all users logged in via OAuth and adds
integration test for the entire OAuth security configuration.