chore(ui): migrate to vite & vitest
This commit also adds minimal url polyfill for tests

NodeJS does not support relative URLs. But our testsuite relies on
relative URLs.
With jest it was not a problem because jest does come with some url
polyfills that fix this problem.
vitest does not setup the same polyfills, to work around this we can
just overwrite the URL constructor and add a URL base to the original
constructor call. This solves the problem and does not require any
complex modifcations/polyfills 🎉