fix: disable platformAutomerge
renovate 36 changed the default setting of `paltformAutomerge` to true.
GitHub does not allow us to block PRs with failing checks from beeing
merged. We can only define a set of "checks that need to run" which is
not suiteable for our monorepo because we can have PRs with changes
only to readmes oder other files that do not trigger pipeline runs...

So to restore the behaviour of pre-36 renovate we'll disable
`platformAutomerge` for now.